Change rating to PositiveSmallIntegerField.
0 to 32767 is enough for a 5 star rating…
https://docs.djangoproject.com/en/5.1/ref/models/fields/#positivesmallintegerfield

…vs -2147483648 to 2147483647
https://docs.djangoproject.com/en/5.1/ref/models/fields/#integerfield

+++ core/models.py
@@ -363,7 +363,7 @@ class BookReading(models.Model):
     start_date = models.DateField()
     end_date = models.DateField(null=True, blank=True)
     finished = models.BooleanField(default=False)
-    rating = models.IntegerField(
+    rating = models.PositiveSmallIntegerField(
         null=True, blank=True, validators=[MinValueValidator(1), MaxValueValidator(5)]
     )
     review = models.TextField(blank=True)
